Solution to Check if a Number is Square of Another

+++ Check if a Number is Square of Another@@6788dd63490637cfc44e1ba4/code.c
@@ -2,8 +2,8 @@
 
 
 int main() {
-    int num1,num2;
-    scanf("%d%d",&num1,&num2);
+    float num1,num2;
+    scanf("%f%f",&num1,&num2);
 
     if(num2==num1*num1){
       printf("Yes");
